Dramatically reduce CPU usage of the NIO thread
Previously this thread effectively infinite looped, as it waited on the
WRITE & CONNECT events for all open sessions, which fired constantly
whilst those sessions were open.

It now waits for CONNECT only until the session is connected, then waits
for READ, and only waits for WRITE for initial data, or if a previous
write didn't complete due to the socket not being writable (generally:
the socket's buffer filling up).

The previous implementation relied implicitly on this constant loop
while sockets were writable to ensure that queued data went out.
Instead, we now have a queue of sessions waiting to send. New sessions
on the queue wake the select thread with a wakeup() message, and after
every select() the thread handles all pending session data in the queue.