Change from custom RC permission to INJECT_EVENTS for easier ADB
INJECT_EVENTS is a permission that allows a process to inject key/touch
events into other apps. Not for 3rd party use, and signature protected
with the system signature. If you have this you have complete power to
automate usage of the HTTP Toolkit app anyway, so it seems sufficient,
and it means that the adb shell user can automate the app with no
user id or permission games at all.