Prompt users to update via play store, if available & necessary
This is intended for users who sideload the app (via the ADB
interceptor) and then continue to use it for a while. These users won't
receive updates at all. Not perfect, but for now we'll nudge them across
onto the play store version in that case. Later the ADB interceptor will
also update installs automatically, to handle users using ADB setup long
term.