Drop non-functional TCP window size checks entirely
Previously we checked the window size, but never incremented or
decremented the count of unacked data, so it was a) confusing
and b) open to the risk that something would increment the
count, thereby blocking the connection entirely.

Now, we just ignore it, effectively using an infinite window
size. If the device is overloaded, we assume our upstream
connections will be applying their own backpressure anyway.
Seems to have worked fine so far, might be worth investigating
if we do see issues on some devices under load.