Ensure the OP_WRITE subscription can't get left on
Unclear how, but it seems that it is still possible for this to happen.
Once it does, that channel spins 100% on the device until it closes, or
until we successfully write something else (and clear up afterwards).
Would be nice to find the root cause, but this solves the immediate
problem safely & clearly.