Ensure the VPN notification remains enabled wherever possible
This used to be unavoidable, but Android 13 has loosened the rules, and
with the recent SDK update notification were no longer appearing by
default. This fixes that, and covers various other scenarios, ensuring
the VPN refuses to start until notifications are activated.

The goal here is partly UX and partly security. This isn't intended to
be used to observe anybody without their knowledge, and this makes that
harder. The resulting UX is also generally better - it's much clearer
when interception is happening, and it's much easier to turn it on &
off.