Separate UDP & TCP session keys to fix VPN bug
Previously, sessions were tracked in one big session table, keyed by
source/destionation ip/port pairs. This sounds great, until you release
that actually UDP & TCP ports are independent, and so it's totally
possible to create a new TCP connection on the same ports that are being
used for an existing UDP connection. Uh oh.

In practice, this meant that SYN packets for TCP connections on existing
UDP ports would attempt to re-ack, rejecting the SYN as a duplicate for
an existing TCP connection. This would then fail, because the session
didn't actually have existing TCP data required to do this. Oops.

We now track them separately, which should avoid that situation
entirely.