Refactor & cleanup logic around post-abort connection shutdown
I don't want to change this behaviour too much as it's all delicate and
poorly understood or tested, but this aborting state is what we hit
after most errors, and triggers a session shutdown, which then results
in downstream RST on any ongoing usage, much like any other silent
connection death.

This change tries to centralize that logic into the session (which is
slowly gaining ownership over session activity) and commonize logic
between the reader, writer & NIO service to link that all together.